pyproject.toml
[project]
name = "mono-pixel"
version = "0.1.0"
description = "a tool that converts your text and fonts into pixel img."
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pillow>=10.0.0", "typer>=0.12.0"]

[project.scripts]
mono-pixel = "mono_pixel.cli:cli"

[dependency-groups]
dev = ["pre-commit>=4.4.0", "ruff>=0.14.5", "ty>=0.0.25", "pytest>=8.0.0"]

## UV
[tool.uv]
package = true

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

## Type checker (ty)
[tool.ty]

[tool.ty.rules]
all = "error"

[tool.ty.analysis]
respect-type-ignore-comments = true

[tool.ty.terminal]
output-format = "concise"


## Linter and formatter (ruff)

[tool.ruff]
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]

[tool.ruff.format]
docstring-code-format = true
